requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiokafka==0.12.0
aiosignal==1.3.2
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
confluent-kafka==2.8.0
dataclasses==0.6
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djongo==1.2.31
dnspython==2.7.0
ecdsa==0.19.0
elastic-transport==8.17.0
elasticsearch==8.17.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
flake8==7.1.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jose==1.0.0
kafka-python==2.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
motor==3.6.1
multidict==6.1.0
mypy-extensions==1.0.0
orjson==3.10.14
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
pyasn1==0.6.1
pycodestyle==2.12.1
pydantic==2.10.5
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.9.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
yarl==1.18.3
